refactored nodes
