+++ components/nodes/filter-node/combo-box.tsx
@@ -1,0 +1,90 @@
+import * as React from 'react';
+import { Check, ChevronsUpDown } from 'lucide-react';
+
+import { cn } from '@/lib/utils';
+import { Button } from '@/components/ui/button';
+import {
+    Command,
+    CommandEmpty,
+    CommandGroup,
+    CommandInput,
+    CommandItem,
+    CommandList,
+} from '@/components/ui/command';
+import {
+    Popover,
+    PopoverContent,
+    PopoverTrigger,
+} from '@/components/ui/popover';
+
+const filters = [
+    {
+        value: 'lowpass',
+        label: 'Low Pass',
+    },
+    {
+        value: 'highpass',
+        label: 'High Pass',
+    },
+];
+
+export default function ComboBox() {
+    const [open, setOpen] = React.useState(false);
+    const [value, setValue] = React.useState('');
+
+    return (
+        <>
+            <Popover open={open} onOpenChange={setOpen}>
+                <PopoverTrigger asChild>
+                    <Button
+                        variant="outline"
+                        role="combobox"
+                        aria-expanded={open}
+                        className="w-[200px] justify-between"
+                    >
+                        {value
+                            ? filters.find(
+                                  (framework) => framework.value === value
+                              )?.label
+                            : 'Select filter...'}
+                        <ChevronsUpDown className="opacity-50" />
+                    </Button>
+                </PopoverTrigger>
+                <PopoverContent className="w-[200px] p-0">
+                    <Command>
+                        <CommandInput placeholder="Search filter..." />
+                        <CommandList>
+                            <CommandEmpty>No filter found.</CommandEmpty>
+                            <CommandGroup>
+                                {filters.map((filter) => (
+                                    <CommandItem
+                                        key={filter.value}
+                                        value={filter.value}
+                                        onSelect={(currentValue) => {
+                                            setValue(
+                                                currentValue === value
+                                                    ? ''
+                                                    : currentValue
+                                            );
+                                            setOpen(false);
+                                        }}
+                                    >
+                                        {filter.label}
+                                        <Check
+                                            className={cn(
+                                                'ml-auto',
+                                                value === filter.value
+                                                    ? 'opacity-100'
+                                                    : 'opacity-0'
+                                            )}
+                                        />
+                                    </CommandItem>
+                                ))}
+                            </CommandGroup>
+                        </CommandList>
+                    </Command>
+                </PopoverContent>
+            </Popover>
+        </>
+    );
+}
