+++ components/nodes/source-node.tsx
@@ -1,10 +1,20 @@
 import { Handle, Position } from '@xyflow/react';
+import {
+    Card,
+    CardHeader,
+    CardContent,
+    CardFooter,
+} from '@/components/ui/card';
 
 export default function SourceNode() {
     return (
-        <div className="px-4 py-2 shadow-md rounded-md bg-white border-2 border-stone-400">
+        <Card>
             <Handle type="source" position={Position.Bottom} />
-            <div className="font-bold">Source</div>
-        </div>
+            <CardHeader></CardHeader>
+            <CardContent>
+                <p>Source</p>
+            </CardContent>
+            <CardFooter></CardFooter>
+        </Card>
     );
 }
