+++ package.json
@@ -32,7 +32,8 @@
     "recharts": "^2.15.1",
     "shadcn-ui": "^0.9.4",
     "tailwind-merge": "^2.6.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "ws": "^8.18.1"
   },
   "devDependencies": {
     "@types/node": "^20",
